% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_sim.R
\name{do_sim}
\alias{do_sim}
\title{Generate multiple ms simulation commands}
\usage{
do_sim(index)
}
\arguments{
\item{index}{simulation number}
}
\description{
This function generates a shell script for all ms coalescent simulations.
}
\examples{
doSim(1)
}
