% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSim.R
\name{doSim}
\alias{doSim}
\title{Generate multiple ms simulation commands}
\usage{
doSim(index)
}
\arguments{
\item{index}{simulation number}
}
\description{
This function generates a shell script for all ms coalescent simulations.
}
\examples{
doSim(1)
}
