% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment_data.R
\name{get_alignment_data}
\alias{get_alignment_data}
\title{Generate single ms simulation command}
\usage{
get_alignment_data(all_data)
}
\arguments{
\item{all_data}{explain}
}
\description{
This function parses input data from alignments and call function to reform them
}
\examples{
get_alignment_data(c("1010", "bananas", "0010", "0110"))
}
