% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rho_data.R
\name{get_rho_data}
\alias{get_rho_data}
\title{Extract rho values from ms simulation data}
\usage{
get_rho_data(all_data)
}
\arguments{
\item{all_data}{explain}
}
\description{
This function finds lines with rho values, extracts, and combines all values into a numeric vector
}
\examples{
get_rho_data(c("cat -r 1.3 20001", "random -r 3.4 20001"))
}
