% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split_index.R
\name{get_split_index}
\alias{get_split_index}
\title{Get indexes used to split data}
\usage{
get_split_index(percent_train, num_sims)
}
\arguments{
\item{percent_train}{Proportion of data used for training set}

\item{num_sims}{Total number of simulations used}
}
\description{
This function figures out the indexes for training, validation, and testing sets, and returns them as a vector for use in splitting functions
}
\examples{
get_split_index(0.8, 1999)
}
