% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_to_array}
\alias{list_to_array}
\title{Convert list of matrices (alignments) to a 3D array}
\usage{
list_to_array(input, dim)
}
\arguments{
\item{input}{A list of matrices.}

\item{dim}{A vector of dimensions of 3D array.}
}
\description{
Convert list of matrices (alignments) to a 3D array
}
