% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_simulate.R
\name{ms_simulate}
\alias{ms_simulate}
\title{Generate single ms simulation command}
\usage{
ms_simulate(morgans_per_bp, ne, filename)
}
\arguments{
\item{morgans_per_bp}{crossover rate per bp per meiosis}

\item{ne}{effective population size}

\item{filename}{output file}
}
\description{
This function generates one ms command
}
\examples{
ms_simulate(morgans_per_bp = 2.018353e-08, ne = 1000, filename = 'sim.txt')
}
