% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msSimulate.R
\name{msSimulate}
\alias{msSimulate}
\title{Generate ms simulations}
\usage{
msSimulate(morgans.per.bp, ne, filename)
}
\arguments{
\item{morgans.per.bp}{explain}

\item{ne}{explain}

\item{filename}{explain}
}
\description{
This function generates a shell script for ms coalescent simulations.
}
\examples{
msSimulate()
}
