% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_single_alignment.R
\name{parse_single_alignment}
\alias{parse_single_alignment}
\title{Generate single alignment from strings}
\usage{
parse_single_alignment(single_alignment)
}
\arguments{
\item{single_alignment}{explain}
}
\description{
This function reforms alignments from separate strings to a matrix
}
\examples{
parse_single_alignment(c("1010", "0010", "0110"))
}
