% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_normalize.R
\name{rho_normalize}
\alias{rho_normalize}
\title{Normalize rho values}
\usage{
rho_normalize(rho_set, mean_center)
}
\arguments{
\item{rho_set}{A vector of rho values used for training, validation, or testing}
}
\description{
This function takes in a vector of rho values and normalizes it by log transforming and centering on a mean value
}
\examples{
rho_normalize(c(7.23, 55.31, 715.52, 2328.45, 423.55), 4.2)
}
