% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_simulate.R
\name{ms_simulate}
\alias{ms_simulate}
\title{Generate single ms simulation command}
\usage{
ms_simulate(morgans.per.bp, ne, filename)
}
\arguments{
\item{morgans.per.bp}{explain}

\item{ne}{explain}

\item{filename}{explain}
}
\description{
This function generates one ms command
}
\examples{
msSimulate(morgans.per.bp = 2.018353e-08, ne = 1000, filename = 'sim.txt')
}
